requirements.txt
discord
python-dotenv
feedparser
babel
python-dateutil